#61 feat : daily 그래프 타임테이블 데이터 연결

+++ core/designsystem/src/main/kotlin/com/titi/app/core/designsystem/model/TdsTimeTableData.kt
@@ -5,6 +5,6 @@ import androidx.compose.runtime.Immutable
 @Immutable
 data class TdsTimeTableData(
     val hour: Int,
-    val start: Long,
-    val end: Long,
+    val start: Int,
+    val end: Int,
 )
